move Observer structs to its own module
